Do not expose the size of recent_keys to Lisp
That prevents from unintentional crashes if the users
modify the variable with setq or if they let-bind it.

Users can still safely modify the lossage limit with the
command `update-lossage-size'.  For convenience, add
a function `lossage-size' to return the current limit.

* src/keyboard.c (lossage_limit): Do not expose it
to Lisp; make it a static variable.
Keep the current Emacs default (300); accept only values >= 100.
(lossage-size): New function; it returns the current size
of recent_keys.
(update-lossage-size):
Rename it from update-lossage-limit (all callers updated);
make it a command.

* doc/emacs/help.texi (Misc Help)
* etc/NEWS
* lisp/cus-start.el
* lisp/edmacro.el
* lisp/help.el:
Update all references with the new name.

* test/src/keyboard-tests.el (keyboard-lossage-limit):
Amend the test.